Make untyped24_pool adjustable in devices.camkes
It seems the value of 12 predates merging ARM VM to CAmkES VM, and we
have no explanation for that particular value. All we know there's 192
MB memory available for misc use.

In hindsight, on VMs with very little guest memory this could have been
a source of confusion; how VM could have worked even if we didn't
provide the guest RAM with "untyped_mmios" attribute?

Signed-off-by: Hannu Lyytinen <hannux@ssrc.tii.ae>