virtio: Use descriptive name for shared memory
The first idea was to use "virtio_shm". But given that all addresses
within virtio descriptors are from driver point of view (in virtio's
device/driver terminology), it would be logical to encode this info to
name of this symbol.

However, using "virtio_driver" would be even more confusing. From virtio
terminology point of view, this would be OK. But usually virtio is used
to provide drivers to guest. Would this mean something related to that
entity providing the drivers, the backend?

Speaking of backend, using terms "backend" and "frontend" seems to be
unambiguous. Hence we tentatively use "virtio_fe". For example:

  const uintptr_t virtio_fe_base;

Signed-off-by: Hannu Lyytinen <hannux@ssrc.tii.ae>