virtio: Require virtioarm library
The only reason we need it is to get VIRTIO_CON_PLAT_INTERRUPT_LINE
from virtioarm/virtio_plat.h.

TODO: Since we have full control of the backend and it's not tied to
any particular hardware platform, perhaps we can define it ourselves
and get rid of this requirement?

In any case, it is probably worth trying to make it possible for
the traditional CAmkES VM virtio and our new modern virtio to co-exist
in the same VM component.

Signed-off-by: Hannu Lyytinen <hannux@ssrc.tii.ae>