Enlarge driver-VM RAM to 512 MB
This is done because for benchmarking purposes we want to cache user-VM
rootfs to tmpfs. At the same time extra space in user-VM's rootfs is
reduced in Yocto configuration, still leaving room for fio benchmark.

Note that similar PR exists here:

https://github.com/tiiuae/tii-sel4-vm/pull/10

Not sure if that extra DMA setting is needed though.

Signed-off-by: Hannu Lyytinen <hannux@ssrc.tii.ae>