intx: Add PCI #INTx support
Traditionally there are four interrupt lines from #INTA to #INTD on the
PCI connector. This code emulates these in software, supporting up to 32
or 64 (depending on word size) devices sharing the same line.

It is also level-sensitive; so far interrupts in CAmkES VM have been
more like edge-triggered. Now the interrupt acknowledge callback
resamples the wired-OR signal from all interrupt sources and re-injects
corresponding IRQ to vCPU if any of the signal sources are still active.

Signed-off-by: Hannu Lyytinen <hannux@ssrc.tii.ae>